Model_Architecture/requirement.txt
pyyaml
torch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1
tensorboard==2.5.0
matplotlib
tqdm
sklearn
librosa==0.8.1